dom/base/test/test_bug453736.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=453736
-->
<head>
  <title>Test for Bug 453736</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=453736">Mozilla Bug 453736</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">
/** Test for Bug 453736 **/
SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  const scriptCreationFuncs = [
    function() { return document.createElement("script"); },
    function() { return document.createElementNS("http://www.w3.org/2000/svg", "script"); }
  ];

  const scriptContainers = ["div", "iframe", "noframes", "noembed"];
  for (var i = 0; i < scriptContainers.length; ++i) {
    for (var func of scriptCreationFuncs) {
      var cont = scriptContainers[i];
      var node = document.createElement(cont);
      document.body.appendChild(node);
      var s = func();
      s.setAttribute("type", "application/javascript");
      s.appendChild(document.createTextNode('window["'+cont+'ScriptRan"] = true'));

      window[cont+"ScriptRan"] = false;
      node.appendChild(s);
      is(window[cont+"ScriptRan"], true,
         "Script created with " + func +" should run when inserting in <"+cont+">");

      window[cont+"ScriptRan"] = false;
      document.body.appendChild(s);
      is(window[cont+"ScriptRan"], false,
         "Script created with " + func + " shouldn't run when moving out of <"+cont+">");

      window[cont+"ScriptRan"] = false;
      document.body.appendChild(s.cloneNode(true));
      is(window[cont+"ScriptRan"], false,
         "Clone of script inside <" + cont + "> created with " + func + " shouldn't run");
    }
  }

  SimpleTest.finish();
});
</script>
</pre>
</body>
</html>
